style(e-commerce shopping): rename "ECommerceShop" folder to "Shop"
